feat: move column sorting related test out of symlinked tests
